requirements.txt
langchain
openai
sentence_transformers
chromadb
langchain-community
python-dotenv